updating files to have everything possible for 1.8.X release
The files now contain most, if not all, options within them to avoid
having to lookup everything in the sample files. It does not contain
some of the comments for explaining what the option is for. For that one
still needs to check the sample file or the manpage for the config.